Tests with L1-Norm as Objective function
The optimisation approach was further developed for two approaches.
In RKoptim it is used to adapt the b in a method with a higher number of stages than the order
In RKoptim-Fallback a method with as many stages as the order is used. If a step does not lead to a positive result the order is reduced and a new b is calculated.

A new testproblem was implemented