Added Notebook to explore the theory that is needed
Propagation Matrix.ipynb 
Notebook to try to formulate the approach as choosing a linear combination of propagation matrices.
Apparently there is no way to get arround the order limitation for implicit methods for the general case

Stabilityfunctions.ipynb
Notebook to try to formulate the behaviour of the approach using linear combinations of the stability functions.
First ideas on how to enshure A-Stability for the implicit methods