Further investigation of fallback approach
According to the results of the previous commit it is clear that enforcing a higher order is not possible for implicit methods. This lead to the use of the fallback approach for implicit methods.
Implicit.ipynb
Added fallback option
Added convergence plot for heat and advection equation
The resulting method is slightly better than the implicit euler while enshuring positifity