Reorganised Files, Added new Testproblem
The implementations for the RK-Integrators where moved to the seperate file. The definition in the notebooks were replaced with imports.

RKimple.py:
File with the implementation of the RK-methods

ADP.ipynb:
New testproblem from https://doi.org/10.1002/pamm.201900209 
This appears to be a interesting testproblem:
Main connclusions: 
- Large b's don't automatically lead to bad results
- For explicit method: It is important to deal with negative values in the stages
- For implicit method: It is important to enshure uniqenes of the roots in the stage equation 

Implicit.ipynb and
RKoptim.ipynb
Minor changes, moved the Implementatition into RKimple.py