Improved the LP solver
The solving of the LP problem was moved into a seperate function to keep it consistant in the explicit and implicit implementation. A option to reduce the number of positifity constraints on the LP-Problem was introduced. 
A interface for an alternate LP solver form scipy was implemented. The options for the cvxpy solvers are now accesible.

Update Description of OrderCondition.py