Tests for explicit Methods, Convergenc
The developement is focused on two main topics: 
- Explicit methods:
	For which Problems and timesteps is the adaptation approach usefull
	Which basis methods are good choiches (in terms of stability and ensured positifity)
	
- Implicit Methods:
	Measure and improve the performance of the implicit solver. In terms of convergenc and speed

utils.py
Added a file with utility functions:
Functions for finding important dt for explicit methods
Function for plotting the convergenc

ADP.ipynb
Further testing on feasible dt for the explicit method
Convergenc plot for the implicit methods

RKimple.py
Minor changes. Mainly to improve usability and for a better handling of errors and numerical problems.