Further explortion for the explicit methods. Improvements of the implementation.
The behavior of explicit basemethod was studied in more detail. For this a new notebook for advection type testproblems was introduced. For a better handling the implementation was improved.

Interesting testproblems are usaully nonlinear. This results in negative values in the solution computete by the ssp104 method that can be potentially fixed with adapting the bs. 
When choosing the baseline method one have to meke sure that the order conditions give some degrees of freedom for choosing a new b.

Advection.ipynb 
New notebook with a linear advection testproblem and a variable coefficent advection testproblem

ADP.ipynb
Tests with more baseline methods. Some new results

OrderConditions.py 
Added the order conditions for 6th and 7th order

RKimple.py
Smal improvements for the returned status informations

RKopitm.ipynb
Added some tests on baseline methods,
Added some testproblems

utils.py
Improvement on automatic search of timesteps
Method for plotting results