Futher improvements on implementation and tests on ADP Problem
utils.py: 
Function to display the status of the Integrator results
Adapted functions for the new implementation in RKimple.py

RKimple.py: Bugfixes, minor improvements

ADP.ipynb
Changed all cells for the new implementation
Main Results: Adapting the b works for all timesteps that are stable, as long as the tol_neg is big enough. ck5 and dp5 lead to positive results for all stable timesteps.